@@ -16,7 +16,8 @@ const bluegrey = const Color(0xFF3E5569);
 const charcoal_light = const Color(0xFF2f363d);
 const charcoal = const Color(0xFF24292E);
 const black = Colors.black;
-const transparent = const Color(0x263E5569);
+const transparent = const Color(0x00ffffff);
+const overlay = const Color.fromRGBO(0, 0, 0, 80);
 
 const ada_pink = const Color(0xFFB94B4B);
 const ada_green = const Color(0xFF3D854A);
@@ -25,8 +26,7 @@ const g_pink_yellow = const LinearGradient(
     colors: [yellow, pink],
     begin: const FractionalOffset(0.4, 0.0),
     end: const FractionalOffset(0.0, 0.5),
-    stops:[0.0,1.0],
-    tileMode: TileMode.clamp
-);
+    stops: [0.0, 1.0],
+    tileMode: TileMode.clamp);
 
 //  GoovingMaterialIcons: https://cdn.materialdesignicons.com/2.6.95/
