@@ -1,33 +1,24 @@
-import 'dart:async';
-import 'dart:io';
-
 import 'package:HackRU/colors.dart';
-import 'package:HackRU/constants.dart';
-import 'package:HackRU/models/loading_indicator.dart';
-import 'package:dart_lcs_client/dart_lcs_client.dart';
+import 'package:HackRU/models/models.dart';
+import 'package:HackRU/screens/newScanner.dart';
+import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:qrcode_reader/qrcode_reader.dart';
 
-// This variable tracks if the user scanned something or just pressed the back
-// button without scanning anything. This is set to false when the scanner returns
-// null to signify that we shouldn't show a progress indicator popup.
-var popup = true;
 const NOT_SCANNED = "NOT SCANNED";
-const instructions = 'Note: Click [Camera Icon] Below to Scan QR Codes!';
 
 @visibleForTesting
 enum Event {
   Check_In,
-  Check_In_No_Delayed,
   Lunch_1,
   Dinner,
   T_Shirt,
   Midnight_Meal,
   Midnight_Surprise,
   Breakfast,
-  Lunch_2
+  Lunch_2,
+  Check_In_No_Delayed,
 }
 
 typedef DemoItemBodyBuilder<T> = Widget Function(DemoItem<T> item);
@@ -103,33 +94,6 @@ class CollapsibleBody extends StatelessWidget {
     final TextTheme textTheme = theme.textTheme;
 
     return Column(children: <Widget>[
-//      const Divider(height: 1.0),
-//      Container(
-//          child: Row(
-//              mainAxisAlignment: MainAxisAlignment.center,
-//              children: <Widget>[
-//            Container(
-//                margin: const EdgeInsets.only(right: 8.0),
-//                child: FlatButton(
-//                    onPressed: onCancel,
-//                    child: const Text('CANCEL',
-//                        style: TextStyle(
-//                            color: charcoal,
-//                            fontSize: 15.0,
-//                            fontWeight: FontWeight.w500)))),
-//            Container(
-//                margin: const EdgeInsets.only(right: 8.0),
-//                child: FlatButton(
-//                  onPressed: onSave,
-//                  child: const Text(
-//                    'SAVE',
-//                    style: TextStyle(
-//                        color: pink,
-//                        fontSize: 15,
-//                        fontWeight: FontWeight.w600),
-//                  ))),
-//          ],),),
-//      const Divider(height: 1.0),
       Container(
           margin: const EdgeInsets.only(left: 24.0, right: 24.0, bottom: 24.0) -
               margin,
@@ -181,16 +145,14 @@ class QRScanner2 extends StatefulWidget {
 
 class _QRScanner2State extends State<QRScanner2> {
   List<DemoItem<dynamic>> _demoItems;
-  Future<String> _message;
   var _isVisible;
 
   @override
   void initState() {
     super.initState();
-    _message = Future<String>.sync(() => instructions);
     _demoItems = <DemoItem<dynamic>>[
       DemoItem<Event>(
-          name: 'Scanning...',
+          name: 'DayOf Event',
           value: Event.Check_In_No_Delayed,
           hint: 'Select Event',
           valueToString: (Event location) => location.toString().split('.')[1],
@@ -236,12 +198,12 @@ class _QRScanner2State extends State<QRScanner2> {
                               },
                               activeColor: pink,
                             ),
-//                            RadioListTile<Event>(value: Event.checkIn, title: const Text('Check-In'), groupValue: field.value, onChanged: field.didChange, activeColor: pink,),
                             RadioListTile<Event>(
                               value: Event.Check_In,
                               title: const Text('Check-In'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -254,6 +216,7 @@ class _QRScanner2State extends State<QRScanner2> {
                               title: const Text('Lunch-1'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -266,6 +229,7 @@ class _QRScanner2State extends State<QRScanner2> {
                               title: const Text('Dinner'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -278,6 +242,7 @@ class _QRScanner2State extends State<QRScanner2> {
                               title: const Text('T-Shirts'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -290,6 +255,7 @@ class _QRScanner2State extends State<QRScanner2> {
                               title: const Text('Midnight-Meal'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -302,6 +268,7 @@ class _QRScanner2State extends State<QRScanner2> {
                               title: const Text('Midnight-Surprise'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -314,6 +281,7 @@ class _QRScanner2State extends State<QRScanner2> {
                               title: const Text('Breakfast'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -326,6 +294,7 @@ class _QRScanner2State extends State<QRScanner2> {
                               title: const Text('Lunch-2'),
                               groupValue: field.value,
                               onChanged: (Event value) {
+                                field.didChange(value);
                                 setState(() {
                                   item.value = value;
                                   close();
@@ -389,47 +358,14 @@ class _QRScanner2State extends State<QRScanner2> {
                       }).toList()),
                 ),
               ),
-              Center(
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: <Widget>[
-                    Image.asset(
-                      'assets/images/hackru_offwhite_logo.png',
-                      width: 200,
-                      height: 200,
-                      semanticLabel: 'Hack RU Logo',
-                    ),
-                    Text(
-                      'Day-Of Events Scanner',
-                      style: TextStyle(
-                          fontSize: 25.0,
-                          fontWeight: FontWeight.w500,
-                          color: off_white),
-                      textAlign: TextAlign.center,
-                    ),
-                    SizedBox(
-                      height: 25.0,
-                    ),
-                    Text(
-                      'Scanning For:',
-                      style: TextStyle(
-                        fontSize: 20.0,
-                        color: off_white,
-                      ),
-                      textAlign: TextAlign.center,
-                    ),
-                    SizedBox(
-                      height: 5.0,
-                    ),
-                    Text(
-                      QRScanner2.event,
-                      style: TextStyle(
-                          fontSize: 40.0,
-                          color: yellow,
-                          fontWeight: FontWeight.w600),
-                      textAlign: TextAlign.center,
-                    ),
-                  ],
+              Container(
+                alignment: Alignment(0.0, 0.0),
+                height: 500.0,
+                child: FlareActor(
+                  'assets/Filip.flr',
+                  alignment: Alignment.center,
+                  fit: BoxFit.contain,
+                  animation: "idle",
                 ),
               ),
             ],
@@ -445,23 +381,11 @@ class _QRScanner2State extends State<QRScanner2> {
                 backgroundColor: yellow,
                 splashColor: pink,
                 onPressed: () async {
-                  print(
-                      "---------------scan button pressed ---------------------");
-                  showDialog(
-                    barrierDismissible: false,
-                    context: context,
-                    builder: (BuildContext context,
-                        {barrierDismissible: false}) {
-                      return new AlertDialog(
-                        backgroundColor: Colors.transparent,
-                        elevation: 0.0,
-                        title: Center(
-                          child: new ColorLoader2(),
-                        ),
-                      );
-                    },
+                  Navigator.of(context).push(
+                    MaterialPageRoute(
+                      builder: (context) => NewScanner(),
+                    ),
                   );
-                  _openQRScanner();
                 },
                 tooltip: 'QRCode Reader',
                 icon: Center(
@@ -485,211 +409,5 @@ class _QRScanner2State extends State<QRScanner2> {
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       floatingActionButtonAnimator: FloatingActionButtonAnimator.scaling,
     );
-  }
-
-  void _scanDialog(String body) async {
-    switch (await showDialog(
-      context: context,
-      builder: (BuildContext context, {barrierDismissible: false}) {
-        return new AlertDialog(
-          backgroundColor: pink,
-          title: Text(body,
-              style: TextStyle(fontSize: 30, color: off_white),
-              textAlign: TextAlign.center),
-          actions: <Widget>[
-            FlatButton(
-              child: Text('OK',
-                  style: TextStyle(fontSize: 20, color: yellow),
-                  textAlign: TextAlign.center),
-              onPressed: () {
-                Navigator.pop(context);
-                _openQRScannerAgain();
-              },
-            ),
-          ],
-        );
-      },
-    )) {
-    }
-  }
-
-  void _openQRScanner() async {
-    var _barcodeString = await new QRCodeReader()
-        .setAutoFocusIntervalInMs(100)
-        .setForceAutoFocus(true)
-        .setTorchEnabled(true)
-        .setHandlePermissions(true)
-        .setExecuteAfterPermissionGranted(true)
-        .scan();
-    print("processing _barcodeString");
-    print(_barcodeString);
-    // If the user didn't scan anything, then _barcodeString will be null. If that is the case, then don't show a progress indicator popup or do anything.
-    popup = _barcodeString != null;
-    if (popup) {
-      showDialog(
-        barrierDismissible: false,
-        context: context,
-        builder: (BuildContext context, {barrierDismissible: false}) {
-          return new AlertDialog(
-            backgroundColor: Colors.transparent,
-            elevation: 0.0,
-            title: Center(
-              child: new ColorLoader2(),
-            ),
-          );
-        },
-      );
-      var message = await _lcsHandle(_barcodeString);
-      // The scanner creates an extraneous item on the Navigator stack. We need to pop it before we can pop the loading indicator.
-      Navigator.pop(context);
-      // Now wew can pop the loading indicator.
-      Navigator.pop(context);
-      if (message != NOT_SCANNED) {
-        _scanDialog(message);
-      }
-    } else {
-      // The scanner creates an extraneous item on the Navigator stack. We need to pop it before we continue.
-      Navigator.pop(context);
-      setState(() {
-        _message = _lcsHandle(_barcodeString);
-      });
-    }
-  }
-
-  void _openQRScannerAgain() async {
-    var _barcodeString = await new QRCodeReader()
-        .setAutoFocusIntervalInMs(100)
-        .setForceAutoFocus(true)
-        .setTorchEnabled(true)
-        .setHandlePermissions(true)
-        .setExecuteAfterPermissionGranted(true)
-        .scan();
-    print("processing _barcodeString");
-    print(_barcodeString);
-    // If the user didn't scan anything, then _barcodeString will be null. If that is the case, then don't show a progress indicator popup or do anything.
-    popup = _barcodeString != null;
-    if (popup) {
-      showDialog(
-        barrierDismissible: false,
-        context: context,
-        builder: (BuildContext context, {barrierDismissible: false}) {
-          return new AlertDialog(
-            backgroundColor: Colors.transparent,
-            elevation: 0.0,
-            title: Center(
-              child: new ColorLoader2(),
-            ),
-          );
-        },
-      );
-      var message = await _lcsHandle(_barcodeString);
-      Navigator.pop(context);
-      if (message != NOT_SCANNED) {
-        _scanDialog(message);
-      }
-    } else {
-      setState(() {
-        _message = _lcsHandle(_barcodeString);
-      });
-    }
-  }
-
-  Future<bool> _scanDialogWarning(String body) async {
-    return showDialog(
-      context: context,
-      builder: (BuildContext context, {barrierDismissible: false}) {
-        return new AlertDialog(
-          backgroundColor: white,
-          title: Text(body,
-              style: TextStyle(fontSize: 30, color: charcoal),
-              textAlign: TextAlign.center),
-          actions: <Widget>[
-            FlatButton(
-                child: Text('OK',
-                    style: TextStyle(fontSize: 20, color: green),
-                    textAlign: TextAlign.center),
-                onPressed: () async {
-                  Navigator.pop(context, true);
-                }),
-            FlatButton(
-                child: Text('CANCEL',
-                    style: TextStyle(fontSize: 20, color: pink),
-                    textAlign: TextAlign.center),
-                onPressed: () {
-                  Navigator.pop(context, false);
-                }),
-          ],
-        );
-      },
-    );
-  }
-
-  Future<String> _lcsHandle(String email) async {
-    String result;
-    print("called lcsHandle with qr:" + email);
-    var user;
-    try {
-      if (email != null) {
-        user = await getUser(DEV_URL, QRScanner2.cred, email);
-        print("scanned user");
-        print(user);
-        if (!user.dayOf.containsKey(QRScanner2.event) ||
-            user.dayOf[QRScanner2.event] == false) {
-          print(user);
-          result = "SCANNED!";
-
-          if (QRScanner2.event == "Check_In_No_Delayed") {
-            if (user.isDelayedEntry() &&
-                !await _scanDialogWarning(
-                    "HACKER IS DELAYED ENTRY! SCAN ANYWAY?")) {
-              return NOT_SCANNED;
-            } else {
-              QRScanner2.event = "Check_In";
-            }
-          }
-
-          if (QRScanner2.event == "Check_In") {
-            printLabel(email, MISC_URL);
-          }
-
-          updateUserDayOf(DEV_URL, QRScanner2.cred, user, QRScanner2.event);
-        } else {
-          result = 'ALREADY SCANNED!';
-          if (QRScanner2.event == "Check_In") {
-            if (await _scanDialogWarning("ALREADY SCANNED! RESCAN?")) {
-              printLabel(email, MISC_URL);
-              result = "SCANNED!";
-            } else {
-              return NOT_SCANNED;
-            }
-          }
-        }
-      } else {
-        print("attempt to scan null");
-      }
-    } on LcsLoginFailed {
-      result = 'LCS LOGIN FAILED!';
-    } on UpdateError {
-      result = 'UPDATE ERROR';
-    } on NoSuchUser {
-      result = 'NO SUCH USER';
-    } on LcsError {
-      result = 'LCS ERROR';
-    } on LabelPrintingError {
-      result = "ERROR PRINTING LABEL";
-    } on ArgumentError catch (e) {
-      result = 'UNEXPECTED ERROR';
-      print(result);
-      print(e);
-      print(e.invalidValue);
-      print(e.message);
-      print(e.name);
-    } on SocketException {
-      result = "NETWORK ERROR";
-    } catch (e) {
-      result = 'UNEXPECTED ERROR';
-      print(e);
-    }
-    return result;
   }
 }
